adressing rerenders when using search
